+++ packages/vfx/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./Emitter"
 export * from "./MeshParticles"
+export { useParticles } from "./ParticlesContext"
 export * from "./timeline"
